Heavy Item Check Open Legend.html
Me again the pest that I am. It seems to be applying the heavy penalty at = to might again, instead of Might +1, I have edited line 193, have a look and see if I'm right on that one? (not sure on the coding. The inventory section looks amazing! I love it.